Consumer Analytics feature code - apply filter on record stats

+++ src/main/java/com/finvendor/common/util/CommonUtil.java
@@ -2,7 +2,7 @@ package com.finvendor.common.util;
 
 public final class CommonUtil {
 
-    public static long calculatePaginationLastPage(String perPageMaxRecords, int totalRecords) {
+    public static long calculatePaginationLastPage(String perPageMaxRecords, long totalRecords) {
         long lastPageNumber;
         long perPageMaxRecordCount = Integer.parseInt(perPageMaxRecords);
         if (perPageMaxRecordCount <= totalRecords) {
